+++ application/backend/src/main/resources/apps/typerefinery/components/layout/header/head-libs.html
@@ -1,0 +1,1 @@
+<script src="/apps/typerefinery/components/layout/header/script.js"></script>
