--- application/backend/src/main/resources/apps/typerefinery/components/layout/navbar/.content.json
@@ -1,6 +1,0 @@
-{
-    "isLayout": true,
-    "group": "Typerefinery - Layout",
-    "sling:resourceType": "ws:Component",
-    "title": "Header (Prime Vue)"
-}
