--- application/backend/src/main/resources/apps/typerefinery/components/layout/navbar/dialog/.content.json
@@ -1,61 +1,0 @@
-{
-    "sling:resourceType": "wcm/dialogs/dialog",
-    "tabs": {
-        "sling:resourceType": "wcm/dialogs/components/tabs",
-        "menuTab": {
-            "sling:resourceType": "wcm/dialogs/components/tab",
-            "label": "Menu",
-            "mobileTitle": {
-                "sling:resourceType": "wcm/dialogs/components/textfield",
-                "name": "mobileMenuTitle",
-                "label": "Mobile Menu Title"
-            },
-            "items": {
-                "name": "navigation/menuItems",
-                "label": "Tabs",
-                "labelText": {
-                    "sling:resourceType": "wcm/dialogs/components/textfield",
-                    "name": "label",
-                    "label": "Label"
-                },
-                "sling:resourceType": "wcm/dialogs/components/multifield",
-                "link": {
-                    "sling:resourceType": "wcm/dialogs/components/pathpicker",
-                    "rootPath": "/content",
-                    "name": "link",
-                    "label": "URL"
-                },
-                "newtab": {
-                    "sling:resourceType": "wcm/dialogs/components/checkbox",
-                    "name": "openInNewTab",
-                    "label": "Open link in a new tab"
-                }
-            }
-        },
-        "ctaTab": {
-            "sling:resourceType": "wcm/dialogs/components/tab",
-            "label": "CTA",
-            "ctaLabel": {
-                "sling:resourceType": "wcm/dialogs/components/textfield",
-                "name": "cta/label",
-                "label": "Display name"
-            },
-            "link": {
-                "sling:resourceType": "wcm/dialogs/components/pathpicker",
-                "rootPath": "/content",
-                "name": "cta/link",
-                "label": "URL"
-            },
-            "openInNewTab": {
-                "sling:resourceType": "wcm/dialogs/components/toggle",
-                "name": "cta/openInNewTab",
-                "label": "Open link in a new tab"
-            },
-            "displayIcon": {
-                "sling:resourceType": "wcm/dialogs/components/toggle",
-                "name": "cta/displayIcon",
-                "label": "Show icon"
-            }
-        }
-    }
-}
