--- application/backend/src/main/resources/apps/typerefinery/components/layout/navbar/head-libs.html
@@ -1,2 +1,0 @@
-<script src="https://unpkg.com/primevue@^3/menubar/menubar.min.js"></script>
-<script src="/apps/typerefinery/components/layout/navbar/script.js"></script>
