--- application/backend/src/main/resources/apps/typerefinery/components/layout/navbar/navbar.html
@@ -1,19 +1,0 @@
-<!--/*
-    Copyright (C) 2022 Typerefinery.io
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-*/-->
-<sly data-sly-use.model="io.typerefinery.websight.models.components.layout.Navbar"
-    data-sly-use.template="./template.html"
-    data-sly-call="${template.navbar @ modelData = model}">
-</sly>
