--- application/backend/src/main/resources/apps/typerefinery/components/layout/navbar/template.html
@@ -1,28 +1,0 @@
-<!--/*
-    Copyright (C) 2022 Typerefinery.io
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-*/-->
-<template data-sly-template.navbar="${ @ modelData }">
-    <p-menubar class="navbar" :model="headerMenuItems">
-        <template #start>
-            <img alt="logo" src="https://user-images.githubusercontent.com/54505967/210701534-576acc09-1e12-46cb-8584-6de76bdea89b.png" height="40" class="mr-2 navbarLogo">
-        </template>
-        <template #end>
-            <button onclick="switchTheme()"id="toggleTheme">
-                theme
-            </button>
-            <p-inputtext placeholder="Search" type="text"></p-inputtext>
-        </template>
-    </p-menubar>
-</template>
