+++ application/backend/src/main/resources/apps/typerefinery/components/structure/page/head-libs.html
@@ -37,11 +37,11 @@
 <!-- Chart JS CDN -->
 <sly data-sly-include="/apps/typerefinery/components/graphs/head-libs.html"></sly>
 
-<!-- Navbar Component -->
-<sly data-sly-include="/apps/typerefinery/components/layout/navbar/head-libs.html"></sly>
-
 <!-- SideBar Component-->
 <sly data-sly-include="/apps/typerefinery/components/layout/sidebar/head-libs.html"></sly>
 
 <!-- Breadcrumb Component-->
 <sly data-sly-include="/apps/typerefinery/components/layout/breadcrumbs/head-libs.html"></sly>
+
+<!-- Header Component-->
+<sly data-sly-include="/apps/typerefinery/components/layout/header/head-libs.html"></sly>
