+++ application/backend/src/main/resources/apps/typerefinery/templates/page/initial/.content.json
@@ -6,7 +6,7 @@
     "rootcontainer": {
       "sling:resourceType": "typerefinery/components/layout/fixedrootcontainer",
       "header": {
-        "sling:resourceType": "typerefinery/components/layout/navbar"
+        "sling:resourceType": "typerefinery/components/layout/header"
       },
       "container": {
         "sling:resourceType": "typerefinery/components/layout/container"
