+++ tests/content/src/main/content/jcr_root/content/typerefinery-showcase/pages/components/content/title/.content.xml
@@ -5,7 +5,7 @@
     xmlns:sling="http://sling.apache.org/jcr/sling/1.0" jcr:primaryType="ws:Page">
     <jcr:content jcr:primaryType="ws:PageContent" jcr:title="Title" sling:resourceType="typerefinery/components/structure/page" ws:template="/apps/typerefinery/templates/page">
         <rootcontainer jcr:primaryType="nt:unstructured" sling:resourceType="typerefinery/components/layout/fixedrootcontainer">
-            <header jcr:primaryType="nt:unstructured" sling:resourceType="typerefinery/components/layout/headerreference"/>
+            <header jcr:primaryType="nt:unstructured" sling:resourceType="typerefinery/components/layout/header"/>
             <maincontainer jcr:primaryType="nt:unstructured" sling:resourceType="typerefinery/components/layout/maincontainer">
                 <pagesection jcr:primaryType="nt:unstructured" sling:resourceType="typerefinery/components/layout/pagesection">
                     <title jcr:primaryType="nt:unstructured" sling:resourceType="typerefinery/components/content/title"/>
@@ -14,7 +14,6 @@
                     <title jcr:primaryType="nt:unstructured" title="test1" sling:resourceType="typerefinery/components/content/title" id="title1"/>
                 </pagesection1>
             </maincontainer>
-            <footer jcr:primaryType="nt:unstructured" sling:resourceType="typerefinery/components/layout/footer"/>
         </rootcontainer>
     </jcr:content>
 </jcr:root>
